+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.5.0 (not yet released)
 
 * Added relationship type order by Relationship Field Type.
+* Added smart group contact data source.
 
 # Version 1.4.0